+++ src/main/java/com/github/theredbrain/manaattributes/ManaAttributes.java
@@ -1,8 +1,10 @@
 package com.github.theredbrain.manaattributes;
 
+import com.github.theredbrain.manaattributes.advancements.criterion.ManaUsingEntityPredicate;
 import com.github.theredbrain.manaattributes.config.ServerConfig;
 import com.github.theredbrain.manaattributes.registry.DataAttachmentRegistry;
 import com.github.theredbrain.manaattributes.registry.EnchantmentEntityEffectRegistry;
+import com.github.theredbrain.manaattributes.registry.EntitySubPredicateTypeRegistry;
 import com.github.theredbrain.manaattributes.world.item.enchantment.AddManaEnchantmentEntityEffect;
 import com.mojang.serialization.MapCodec;
 import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
@@ -25,6 +27,8 @@ public class ManaAttributes implements ModInitializer {
 	public static Holder<Attribute> MANA_TICK_THRESHOLD;
 	public static Holder<Attribute> RESERVED_MANA;
 
+	public static MapCodec<ManaUsingEntityPredicate> MANA_USING_ENTITY_PREDICATE;
+
 	public static MapCodec<AddManaEnchantmentEntityEffect> ADD_MANA;
 
 	@Override
@@ -34,6 +38,7 @@ public class ManaAttributes implements ModInitializer {
 
 		DataAttachmentRegistry.init();
 		EnchantmentEntityEffectRegistry.init();
+		EntitySubPredicateTypeRegistry.init();
 	}
 
 	public static void info(String message) {